Moved BalanceApi trait
